Fix typo in FE_Nedelec
git-svn-id: https://svn.dealii.org/trunk@22606 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/source/fe/fe_nedelec.cc
@@ -2784,7 +2784,7 @@ FE_Nedelec<dim>::get_subface_interpolation_matrix(
                     
                     for (unsigned int i = 0; i <= deg; ++i)
                       {
-                        const unsigned L_i = weight
+                        const double L_i = weight
                                            * legendre_polynomials[i].value
                                              (quadrature_points[q_point] (0));
